+++ webapp/app/scripts/openrole/LoaderService.js
@@ -1,7 +1,7 @@
 "use strict";
 
 angular.module('openrole')
-  .factory('loaderService', ['$rootScope','$http','$translate','alertService', function($rootScope, $http, $translate, alertService) {
+  .factory('loaderService', ['$rootScope','$http','$translate','$location','alertService', function($rootScope, $http, $translate, $location, alertService) {
     $rootScope.loading = $translate.instant('LOADING'); // init loading text
 
     var internalSetLoadingReady = function(newState, msg) {
@@ -12,11 +12,33 @@ angular.module('openrole')
       }
     };
 
-    var internalImageLoader = function(imageRecords, imageLoaded, index, callbackFunctionAfterLoad){
+
+    var createInternalPDF = function(scope) {
+      try {
+        var tmplName = eval(scope.openrole.pdftemplate);
+        // TODO: deepcopy --> angular.copy doesnt seem to work
+        var doc = tmplName(scope.openrole, scope.imageLoaded, $translate, alertService, {});
+        if (!angular.isDefined(doc)) {
+          alertService.danger("Template did not return the document.");
+        } else {
+          if ($location.$$host == 'localhost') {
+            // for development: open in new window (not working in IE)
+            doc.output('dataurlnewwindow');
+          } else {
+            // for production: download file
+            doc.save("openrole.pdf");
+          }
+        }
+      } catch (e) {
+        alertService.danger(e);
+      }
+    };
+
+    var internalImageLoader = function(imageRecords, imageLoaded, index, scope){
       //bind load event
       if( index >= imageRecords.length ){
         internalSetLoadingReady(false, $translate.instant('GENERATING_PDF'));
-        callbackFunctionAfterLoad();
+        createInternalPDF(scope);
         internalSetLoadingReady(true);
       } else {
         internalSetLoadingReady(false, $translate.instant('LOADING') + " "+(index+1)+"/"+imageRecords.length);
@@ -26,7 +48,7 @@ angular.module('openrole')
             console.log("Loaded image "+index+" "+imageRecords[index]);
             // convert to dataurl, because jsPDF displays dataURLs nicer
             imageLoaded[index] = data; //"data:image/jpeg;base64," + Base64.encodeBinary(data);
-            internalImageLoader(imageRecords, imageLoaded, index + 1, callbackFunctionAfterLoad);
+            internalImageLoader(imageRecords, imageLoaded, index + 1, scope);
           })
           .error(function(data, status, headers, config){
             alertService.danger(status+": "+$translate.instant('ERROR_LOAD_IMG')+" "+data);
@@ -37,8 +59,8 @@ angular.module('openrole')
     };
 
     return {
-      loadImages: function(imageRecords, imageLoaded, callbackFunctionAfterLoad){
-        internalImageLoader(imageRecords, imageLoaded, imageLoaded.length, callbackFunctionAfterLoad);
+      loadImagesAndGeneratePDF: function(scope){
+        internalImageLoader(scope.imageToLoad, scope.imageLoaded, scope.imageLoaded.length, scope);
       },
       setLoadingReady: function() {
         internalSetLoadingReady(true);